fixed #2: inconsistent references weather.py
